move consts & tests around
